Updated category and individual classes to take in doubles for grades

+++ Category.h
@@ -7,12 +7,12 @@ class Category{
 private:
     std::string category;
     std::string command;
-    std::vector<std::vector<int>> grades;
+    std::vector<std::vector<double>> grades;
     int DetermineCategory(std::string category);
 public:
     friend void TestCategoryDetermineCategory();
-    Category(std::string category, std::string command, std::vector<std::vector<int>> grades);
-    std::string AllGrades(std::string category, std::vector<std::vector<int>> grades);
-    double Total(std::string category, std::vector<std::vector<int>> grades);
+    Category(std::string category, std::string command, std::vector<std::vector<double>> grades);
+    std::string AllGrades(std::string category, std::vector<std::vector<double>> grades);
+    double Total(std::string category, std::vector<std::vector<double>> grades);
 };
 #endif //CSC212_GRADEBOOK_CATEGORY_H
