+++ Individual.h
@@ -11,7 +11,7 @@ class Individual
 private:
     std::string category;
     std::string command; // Deliverable
-    std::vector<std::vector<int>> grades; 
+    std::vector<std::vector<double>> grades;
     std::vector<std::vector<std::string>> assignment_names; // Vector of deliverable names
     int DetermineCategory(std::string category); // Implimentation from category class
     // Find index of given deliverable
@@ -19,9 +19,9 @@ private:
 
 public:
     friend void TestIndivDetermineCategory();
-    Individual(std::string category, std::string command, std::vector<std::vector<int>> grades, std::vector<std::vector<std::string>> assignment_names);
+    Individual(std::string category, std::string command, std::vector<std::vector<double>> grades, std::vector<std::vector<std::string>> assignment_names);
     // Return grade at deliverable index
-    int GetGrade(std::string category, std::string command, std::vector<std::vector<int>> grades);
+    double GetGrade(std::string category, std::string command, std::vector<std::vector<double>> grades);
 };
 
 #endif // CSC212_GRADEBOOK_INDIVIDUAL_H
